Added Game class prototype, renamed game object to StandardGame
This indicates waypoint to the new architecture, where Game will not be singletonic.
Singletonic Game was only a way to fasten prototype development, it was not intended
to stay.

Possibility to instantiate Game's subtypes gives more flexibility and allows to create
different game types, such as multiplayer over the network. Also, this allows to keep
simultanously different games taking place on one device, but I doubt this is the path
that should be taken.